requirements.txt
pinecone-client[grpc]
openai
IPython